requirements.txt
streamlit
python-dotenv
langchain-community
langchain-huggingface
langchain-pinecone
langchain
langchain-groq
pinecone-client
sentence-transformers
crewai